smart_agri_models/disease_detection/requirements.txt
torch
pillow
torchvision
fastapi
uvicorn 